Remove previous em mouse

+++ src/input/mouse.cpp
@@ -3,7 +3,7 @@
 namespace dxna::input {
 	int Mouse::X{ 0 };
 	int Mouse::Y{ 0 };
-	int Mouse::Wheel{ 0 };
+	int64_t Mouse::Wheel{ 0 };
 	ButtonState Mouse::Left{ ButtonState::Released };
 	ButtonState Mouse::Right{ ButtonState::Released };
 	ButtonState Mouse::Middle{ ButtonState::Released };
